+++ apps/stories/migrations/0009_auto_20150224_1014.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import models, migrations
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('stories', '0008_story_user_type'),
+    ]
+
+    operations = [
+        migrations.RunSQL('ALTER TABLE stories_question ADD COLUMN user_type_id INTEGER REFERENCES stories_usertype'),
+    ]