+++ assets/static/js/klp/klp.codekit.js
@@ -1,7 +1,0 @@
-// @codekit-prepend "klp.volunteer-initial.js"
-// @codekit-prepend "klp.volunteer-map.js"
-// @codekit-prepend "klp.volunteer.js"
-
-$(document).ready(function(){
-
-});