+++ assets/static/js/school/share_story.js
@@ -20,7 +20,16 @@
     };
 
     function doPostFormRender() {
+        initializeUser();
         $('[name=image]').imagePreview();
+        //datepicker 
+        $(".js-input-date").pickadate({
+            format: 'yyyy-mm-dd',
+            formatSubmit: 'yyyy-mm-dd'
+        });
+        $('#comments_id').maxChars({
+            'max': 2000
+        });
         $('#sysForm').submit(function(e) {
             if (e) {
                 e.preventDefault();
@@ -31,6 +40,11 @@
             _(dataArray).each(function(a) {
                 dataObj[a.name] = a.value;
             });
+            var commentsChars = $('#comments_id').val().length;
+            if (commentsChars > 2000) {
+                klp.utils.alertMessage("Comments field longer than allowed length. Please fix and try again.", "error");
+                return;
+            }
             dataObj['images'] = getImagesData();
             var postURL = "stories/" + SCHOOL_ID;
             // console.log("data obj", dataObj);
@@ -48,9 +62,19 @@
         });
     }
 
+    function initializeUser() {
+        if (klp.auth.getToken()) {
+            klp.api.authDo("users/profile").done(function(data) {
+                $('#name_id').val(data.first_name + " " + data.last_name);
+                $('#email_id').val(data.email);
+                $('#telephone').val(data.mobile_no);
+            });
+        }
+    }
+
     function getImagesData() {
         var images = [];
-        $('.imagePreview').each(function() {
+        $('.js-image-preview').each(function() {
             var src = $(this).attr('src');
             if (src) {
                 images.push(src);